requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.5.0
djangorestframework-simplejwt==5.3.1
